Update dexamarin.py
if-else statement so that `os.remove` does not fail. If it is a directory, it will instead use `shutil.rmtree`. This could be done in the exception clause but as it just captures `PermissionError`exceptions I preferred to let it as it was. Further optimizations or checks might be needed.